requirements.txt
asgiref==3.3.1
astroid==2.5.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.7
django-adminlte-3==0.1.6
django-crispy-forms==1.11.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
gunicorn==20.1.0
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.1.1
psycopg2==2.8.6
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
whitenoise==5.2.0
wrapt==1.12.1
